Map published date as best, published_online as fallback
http://www.rioxx.net/schema/v2.0/rioxx/rioxxterms_.html#publication_date
refers to the publication_date as:
the publication date of the resource in the form in which it would be
cited. This element is not used in a RIOXX context but allows for a
RIOXX record to become a reasonable bibliographic record for the
resource. This is a free-text field.